Complete line drawing re-write using Vectrosity.
Down to one draw call!

Also, performance on updating the lines should be faster:

The BuildLines script now builds a Dictionary, keyed by the VectorLine
objects which make up the character. The contents of each dictionary
entry is a List<Transform>. This cache makes it quicker to find the
lines than the last version.